% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clust_dist}
\alias{clust_dist}
\title{Distance mutation clusters}
\usage{
clust_dist(vr, rand_df, ce_cutoff = 1, method = "fdr", n = 1,
  dist_cutoff = NULL, split_factor = NULL)
}
\arguments{
\item{vr}{A VRanges object with multiple samples (or just one)}

\item{rand_df}{A randommut dataframe with the randomized values}

\item{ce_cutoff}{number of nucleotides that are considered to classify complex events}

\item{method}{fdr for local (per mutation) and FDR for tail-based (sample - wise) FDR}

\item{n}{Number of mutations enclosed in the IMD calculation. N = 1 for pairs of mutations.}

\item{dist_cutoff}{(FDR only) distance cutoff}

\item{split_factor}{(optional) By default it will split the VR by sample, however, to get a stratified boosting different factors can be provided. A vector equal size as the VR with the groups to split.}
}
\value{
A VR object with an extended metadata column with FDR or fdr values.
}
\description{
Detect clusters of mutaion based on the chromosomic distance or IMD.
}
