% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_helpers.R
\name{clustmut_internal_check_dependencies}
\alias{clustmut_internal_check_dependencies}
\title{Check dependencies}
\usage{
clustmut_internal_check_dependencies(deps)
}
\arguments{
\item{deps}{character vector with package names}
}
\description{
Check dependencies
}
