% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_helpers.R
\name{clustmut_internal_return_version}
\alias{clustmut_internal_return_version}
\title{Return version of code}
\usage{
clustmut_internal_return_version(version, quit)
}
\arguments{
\item{version}{TRUE if you want to return the version of the code}

\item{quit}{TRUE if quitting after the version has been rinted in the CLI}
}
\description{
Return version of code
}
