% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{compute_densityfdr}
\alias{compute_densityfdr}
\title{Gaussian density local fdr estimator}
\usage{
compute_densityfdr(obs, null, monotonic = TRUE, alternative = c("left",
  "right"))
}
\arguments{
\item{obs}{nonnull distribution of values}

\item{null}{expected or null distribution}

\item{monotonic}{if fdr should be monotonic}

\item{alternative}{left if fdr should be ascending, right if decending}
}
\value{
local fdr vector
}
\description{
It computes the local fdr estimate based on the density estimates from a nonnull and a null distribution.
}
