% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{compute_distance_vr}
\alias{compute_distance_vr}
\title{Find IMD per mutation k pairs}
\usage{
compute_distance_vr(vr, enclosing)
}
\arguments{
\item{vr}{a VRanges object}

\item{enclosing}{Number of mutations between pairs}
}
\value{
a VRanges object with a mdist column
}
\description{
In a VR object, finds mutation pairs that enclose n number of mutations. It
only selects the nearest pair.
As a general way to handle this I would use nearesDistance if k=1.
}
