% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{compute_distances}
\alias{compute_distances}
\title{Get mutation distance}
\usage{
compute_distances(x, k = 1, use = dplyr::lag)
}
\arguments{
\item{k}{}
}
\description{
Get the distance between a particular mutaion and the next or to the previous.
}
