% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{compute_distances_splited}
\alias{compute_distances_splited}
\title{Compute Distances in splited GR}
\usage{
compute_distances_splited(gr, f, no_cores = NULL)
}
\arguments{
\item{gr}{Genomic Ranges object.}

\item{f}{factor to group rows, if more than one group is required you should input a list containing both factors.}

\item{no_cores}{if null normal execution, if integer, parallel mode is triggered with the number of cores specified.}
}
\description{
Compute minimum distance between adjacent mutation spliting by a factor. This
factor can be either the seqnames (chromosomes) and/or samples.
}
