% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{compute_m_distance}
\alias{compute_m_distance}
\title{Get minimum adjacent distance}
\usage{
compute_m_distance(x, k = 1, use = min)
}
\arguments{
\item{x}{vector of positions (doesn't check same chromosome)}

\item{k}{adjacent positions (k = 1 by default)}

\item{use}{function to select appropiate adjacent mutation (min by default)}
}
\value{

}
\description{
Wrapper of compute_distances
Get the distance between each mutation and the closest adjacent one.
}
