% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_alternative_methods.R
\name{custom_basic_clustering}
\alias{custom_basic_clustering}
\title{Compute cluster based in custom thresholds}
\usage{
custom_basic_clustering(vr, IMD, nmuts, event_categories, nearest = TRUE)
}
\arguments{
\item{vr}{a VRanges object}

\item{IMD}{the inter-mutational distance to select mutations}

\item{nmuts}{the number of mutations per cluster requiered}

\item{event_categories}{a named vactor with the minimum number of mutations}
}
\value{

}
\description{
Compute cluster based in custom thresholds
}
