% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{detect_events}
\alias{detect_events}
\title{Detect events from fdr vectors}
\usage{
detect_events(x, sig_cutoff, event_categories)
}
\arguments{
\item{x}{a numeric vector with lfdr values}

\item{sig_cutoff}{the significance cutoff}

\item{event_categories}{a named vector with number of the mutations in the event as values and names of the events as names.}
}
\value{
Returns a list with events and length of the stretch. Both with same total length as the fdr vector provided.
}
\description{
Detect events from fdr vectors
}
\examples{

fdr_test =c(runif(10,min = 0.21),
c(.19,.1,.05),
runif(10,min=0.21),
c(.1,.1,.1,.1,.1))
categories_test = c("kataegis" = 5,"omikli" = 2)

detect_events(x = fdr_test,
  sig_cutoff = 0.2,
  event_categories =categories_test )
}
