% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{find_all_runs}
\alias{find_all_runs}
\title{Find all mutation runs}
\usage{
find_all_runs(vr, IMD)
}
\arguments{
\item{vr}{a VRanges object}

\item{IMD}{The IMD threshold (equal or smaller)}
}
\value{
A data frame with 3 columns, a from value which returns the original idx
of the run-pair, a to column with the final idx of the run and a all TRUE
mask column which is used for validation.
}
\description{
It search all the possible run-pairs according to a minimum inter-mutational
distance (IMD) in the vr objects and selects the ones
which have at least a given number of mutations.
}
\details{
A run-pair is a set of consequtive pairs which each consists in less than
a given IMD.
}
