% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{find_pairs_VR}
\alias{find_pairs_VR}
\title{Find pairs within a IMD range}
\usage{
find_pairs_VR(vr, win_length)
}
\arguments{
\item{vr}{a VRanges object}

\item{win_length}{the IMD in which pairs will be found}
}
\value{

}
\description{
Find pairs within a IMD range
}
