% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf.R
\name{load_vaf_fdr}
\alias{load_vaf_fdr}
\title{We can see an enrichment of the low vaflr section.
We transform the values using the break_points previously defined
We can see an enrichment of the low vaflr section.
We transform the values using the break_points previously defined}
\usage{
load_vaf_fdr(dat_gr, clust_mask, break_points)
}
\description{
We can see an enrichment of the low vaflr section.
We transform the values using the break_points previously defined
We can see an enrichment of the low vaflr section.
We transform the values using the break_points previously defined
}
