% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf.R
\name{local_vaflr_fdr}
\alias{local_vaflr_fdr}
\title{Detect clusters using per mutation VAF fdr}
\usage{
local_vaflr_fdr(vr, simulation_size_input, pairs_size = 10000)
}
\arguments{
\item{vr}{A VRanges object with somatic mutations}

\item{simulation_size_input}{The number of simulated VAFs to generate for the expected and idealized distribution}

\item{pairs_size}{The lenght (in bp) used to find possible pairs of mutations.}
}
\value{
A VRanges object with added columns (n and vaf_fdr_median)
}
\description{
Detect clusters using per mutation VAF fdr
}
