% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mask_complex_events}
\alias{mask_complex_events}
\title{Mask complex events and unique mutations in a chromosome}
\usage{
mask_complex_events(gr, cutoff = 1)
}
\arguments{
\item{gr}{a VRanges object (unisample)}

\item{cutoff}{minimum distance to consider as a complex event}
}
\value{

}
\description{
Mask complex events and unique mutations in a chromosome
}
