% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_alternative_methods.R
\name{roberts_clusters}
\alias{roberts_clusters}
\title{Cluster mutations Roberts 2012}
\usage{
roberts_clusters(vr, delta = 10000, ce_cutoff = 10, G_percent = 1,
  pval_cutoff = 1e-04, dbSNP)
}
\arguments{
\item{vr}{A VRanges object}

\item{delta}{The maximum window size to consider mutations candidates for clustering}

\item{ce_cutoff}{Complex events cutoff, only last one will be kept}

\item{G_percent}{Percentatge of the human genome used}

\item{pval_cutoff}{P value to decide}

\item{dbSNP}{a SNPLoc object that contains the appropiate dbSNP release}
}
\description{
This functions implements the method in Roberts, S. A. et al. Molecular Cell (2012) to call clusterd mutations. Regions with local higher hypermutation than expected.
}
\details{
WARNING: This method considers the MR constant in each sample.
}
