% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_alternative_methods.R
\name{roberts_pvalue}
\alias{roberts_pvalue}
\title{Compute the negative binomial pvalue}
\usage{
roberts_pvalue(x, k, p)
}
\arguments{
\item{x}{number of nucleotides in the cluster}

\item{k}{number of mutations in the cluster}

\item{p}{mutation probability}
}
\value{

}
\description{
Compute the negative binomial pvalue
}
