% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{to_monotonic}
\alias{to_monotonic}
\title{Transform a vector to monotonic}
\usage{
to_monotonic(x, max = 1, ascending = TRUE, ...)
}
\arguments{
\item{x}{the vector}

\item{max}{maximum value allowed}

\item{ascending}{TRUE if vector should go from 0 to max, False if max to 0}

\item{...}{}
}
\description{
Transform a vector to monotonic
}
