% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{find_all_pairs}
\alias{find_all_pairs}
\title{Find pairs within a IMD range}
\usage{
find_all_pairs(vr, win_length)
}
\arguments{
\item{vr}{a VRanges object}

\item{win_length}{the IMD in which pairs will be found}
}
\value{

}
\description{
Find pairs within a IMD range
}
